refactor: move starred API schema to shared schema.ts
Consolidates starred pattern types with other shared schemas for consistency and type safety.

## Changes

**Schema Consolidation** (`src/schema.ts`)
- Add `StarredPattern` schema: ruleset_name, ruleset_hex, seed, seed_type, seed_percentage
- Add `StarredResponse` schema: ok boolean + nullable pattern
- Export TypeScript types: `StarredPattern`, `StarredResponse`
- Reuse existing `SeedType` enum for consistency
- Match snake_case naming convention from database/API layer

**API Wrapper Refactor** (`src/api/starred.ts`)
- Remove duplicate schema definitions
- Import types from shared `schema.ts`
- Follow same pattern as `save.ts` and `leaderboard.ts`:
  - Import type and validator from schema
  - Schema validation with `.parse()`
  - Consistent error handling
  - Consistent logging format
- Improve JSDoc with detailed parameter description
- Log detailed success info (ruleset name, hex, seed, seed type)

## Benefits

- **Type Safety**: All API schemas in one place
- **Consistency**: Same pattern across all API wrappers
- **Maintainability**: Single source of truth for types
- **DRY**: No duplicate schema definitions
- **Validation**: Zod validates all responses at runtime

## Pattern Match

Now matches existing API wrapper pattern:
- `src/api/save.ts` → imports `SaveResponse` from schema
- `src/api/leaderboard.ts` → imports `LeaderboardResponse` from schema
- `src/api/starred.ts` → imports `StarredResponse` from schema ✅
