+++ src/schema.ts
@@ -243,6 +243,23 @@ export const LeaderboardResponse = z.object({
 })
 
 // ============================================================================
+// Starred pattern types
+// ============================================================================
+
+export const StarredPattern = z.object({
+  ruleset_name: z.string(),
+  ruleset_hex: z.string().regex(/^[0-9a-f]{35}$/i),
+  seed: z.number().int(),
+  seed_type: SeedType,
+  seed_percentage: z.number().nullable(),
+})
+
+export const StarredResponse = z.object({
+  ok: z.boolean(),
+  pattern: StarredPattern.nullable(),
+})
+
+// ============================================================================
 // Type exports
 // ============================================================================
 export type RunSubmission = z.infer<typeof RunSubmission>
@@ -266,3 +283,5 @@ export type Ruleset = z.infer<typeof Ruleset>
 export type C4Ruleset = z.infer<typeof C4Ruleset>
 export type LeaderboardEntry = z.infer<typeof LeaderboardEntry>
 export type LeaderboardResponse = z.infer<typeof LeaderboardResponse>
+export type StarredPattern = z.infer<typeof StarredPattern>
+export type StarredResponse = z.infer<typeof StarredResponse>
